refactor: consolidate approval review cards into shared base component
