+++ frontend/src/components/ChangeRequests/PreAwardReviewCard/PreAwardReviewCard.jsx
@@ -1,113 +1,35 @@
 import PropTypes from "prop-types";
-import { useNavigate } from "react-router-dom";
-import { formatDateToMonthDayYear } from "../../../helpers/utils";
-import { useGetAgreementName } from "../../../hooks/lookup.hooks";
-import useGetUserFullNameFromId from "../../../hooks/user.hooks";
-import TermTag from "../TermTag";
-import { faClock } from "@fortawesome/free-regular-svg-icons";
-import { faEye } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import ApprovalFlowReviewCard from "../ApprovalFlowReviewCard";
 
 /**
  * PreAwardReviewCard component for displaying pending pre-award approval requests
+ * from Division Directors.
  *
- * NOTE: This component does NOT use the shared ReviewCard base component because it serves a different
- * purpose than change request cards. ReviewCard is designed for inline approve/reject actions on
- * change requests, while this card (like BudgetTeamRequisitionReviewCard) is an informational card
- * that links to a detail/approval page.
+ * This is a thin wrapper around ApprovalFlowReviewCard that provides pre-award-specific
+ * configuration (heading, navigation path, button text).
  *
  * @component
  * @param {Object} props - Properties passed to component
  * @param {number} props.agreementId - The ID of the agreement
  * @param {number} props.requestorId - The ID of the user who requested approval
  * @param {string} props.requestDate - The date of the approval request
- * @param {string} [props.requestorNotes] - Notes from the requestor
+ * @param {number} props.executingBliCount - Count of executing budget line items
+ * @param {number} props.executingTotal - Total amount of executing budget line items
+ * @param {string} [props.obligateByDate] - Earliest obligate-by date from executing BLIs
+ * @param {number} props.agreementTotal - Total agreement amount
  * @param {boolean} [props.isCondensed=false] - Whether the card is condensed
  * @param {boolean} [props.forceHover=false] - Whether to force hover state
  * @returns {JSX.Element} - The rendered component
  */
-function PreAwardReviewCard({
-    agreementId,
-    requestorId,
-    requestDate,
-    requestorNotes,
-    isCondensed = false,
-    forceHover = false
-}) {
-    const navigate = useNavigate();
-    const agreementName = useGetAgreementName(agreementId);
-    const requestorName = useGetUserFullNameFromId(requestorId);
-
-    const handleViewClick = (e) => {
-        e.preventDefault();
-        navigate(`/agreements/${agreementId}/review-pre-award`);
-    };
-
+function PreAwardReviewCard(props) {
     return (
-        <div
-            className={`width-full flex-column padding-2 margin-top-4 bg-white hover:bg-base-lightest border-2px radius-lg ${
-                forceHover ? "bg-base-lightest border-base-lighter" : "border-base-light hover:border-base-lighter"
-            }`}
-            data-cy="pre-award-review-card"
-            style={{ minHeight: "8.375rem" }}
-        >
-            {!isCondensed && (
-                <header className="display-flex flex-justify">
-                    <div className="display-flex">
-                        <h2 className="margin-0 font-sans-sm">
-                            Pre-Award Approval
-                            <br />
-                            Request
-                        </h2>
-                        <dl className="font-12px margin-0 margin-left-4">
-                            <dt className="margin-0 text-base-dark">Agreement</dt>
-                            <dd className="margin-0">{agreementName}</dd>
-                        </dl>
-                    </div>
-                </header>
-            )}
-            <section
-                className="display-flex flex-justify margin-y-1"
-                style={{ maxWidth: "50rem" }}
-            >
-                <dl className="font-12px grid-col-2">
-                    <dt className="text-base-dark">Requested By</dt>
-                    <dd className="margin-0">{requestorName || "Unknown"}</dd>
-                </dl>
-                {requestorNotes && (
-                    <TermTag
-                        label="Notes"
-                        value={requestorNotes}
-                        className="grid-col-6"
-                    />
-                )}
-            </section>
-            <footer className="font-12px display-flex flex-justify flex-align-center">
-                <div className="text-base-dark display-flex flex-align-center">
-                    <FontAwesomeIcon
-                        icon={faClock}
-                        className="height-2 width-2 margin-right-1"
-                        aria-hidden="true"
-                    />
-                    {formatDateToMonthDayYear(requestDate)}
-                </div>
-                {!isCondensed && (
-                    <button
-                        type="button"
-                        onClick={handleViewClick}
-                        className="usa-button--unstyled text-primary font-12px cursor-pointer"
-                        data-cy="review-pre-award"
-                    >
-                        Review Request
-                        <FontAwesomeIcon
-                            icon={faEye}
-                            size="3x"
-                            className="height-2 width-2 margin-left-1"
-                        />
-                    </button>
-                )}
-            </footer>
-        </div>
+        <ApprovalFlowReviewCard
+            {...props}
+            headingText="Pre-Award"
+            navigationPath="review-pre-award"
+            dataCyPrefix="pre-award-review-card"
+            buttonText="Review Agreement"
+        />
     );
 }
 
@@ -115,7 +37,10 @@ PreAwardReviewCard.propTypes = {
     agreementId: PropTypes.number.isRequired,
     requestorId: PropTypes.number.isRequired,
     requestDate: PropTypes.string.isRequired,
-    requestorNotes: PropTypes.string,
+    executingBliCount: PropTypes.number.isRequired,
+    executingTotal: PropTypes.number.isRequired,
+    obligateByDate: PropTypes.string,
+    agreementTotal: PropTypes.number.isRequired,
     isCondensed: PropTypes.bool,
     forceHover: PropTypes.bool
 };
